Extract heuristics into dedicated class
